requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
mysqlclient==2.2.4
Werkzeug==3.1.2
WTForms==3.1.2
pdfplumber==0.7.7
